fix(gui): guarantee recorder hotkey-resume on window close (#213)
Review follow-ups on the armed-recorder suspension:

- Add a guaranteed disarm net in SettingsWindowController.
  windowWillClose. Disarm was edge-triggered off the SwiftUI
  field lifecycle, but the Settings window is retained
  (isReleasedWhenClosed = false) and only ordered out, so a
  close route that never delivers a local mouse-down or app
  deactivation could strand the hotkeys suspended until the next
  reload. setRecorderArmed(false) is idempotent, so it is a
  no-op when nothing is armed.
- Comment the load-bearing release-before-commit ordering in
  KeyRecorderField.finish() — resume must precede live-apply so
  its activationFailures read (and the caption) stay honest.
- Split KeyComboTests out of KeybindingTests.swift; adding the
  suspend/resume tests pushed it past the 350-line ceiling.
